+++ bindings/wysiwyg-ffi/src/ffi_composer_state.rs
@@ -1,5 +1,5 @@
 pub struct ComposerState {
     pub html: Vec<u16>,
-    pub start: u64,
-    pub end: u64,
+    pub start: u32,
+    pub end: u32,
 }
